backend/requirements.txt
fastapi
uvicorn
pydantic==2.7.4
python-dotenv
langchain
langchain-core
langchain-community
langchain-text-splitters
openai
chromadb
httpx